docs: make code snippet in README.md easier to copy/paste
The shell command to build the docs is quite long,
making it tricky to copy/paste when viewing the rendered
Markdown on GitHub.

This moves the command into a `build.sh` script so that:

- It is easy to copy/paste from the rendered Markdown.
- We can update `./build.sh` going forward, as needed, without
  users having to learn a new command.

Also amended the docs to include a tip for running a local
webserver, as I personally can never remember this offhand.